[buildarmpkg] use bind mount instead of cp
Using mount -o bind instead of copying the PKGBUILD into the rootfs has multiple pros.
One of which is that we get the updated PKGBUILD's out to the host, if it has a pkgver() function.
We also get the resulting package out in that folder.

Signed-off By: Dan Johansen <strit@manjaro.org>