[buildarmoem] disable resolved on PinePhone
systemd-resolved gets auto enabled on Arch systems. On regular devices we disable it with the OEM setup,
but since the Phone and tablet does not have an OEM setup, we have to do it in the tools.
So we edit the preset file to disable systemd-resolved by default on those devices.
Signed-off-by: Dan Johansen <strit@manjaro.org>