Fix permissions on CI
Since Ubuntu, which we use in the CI, does not have the polkitd group,
it results in some permissions from our overlays to not be correct
when building on the CI.
So instead of using group names, use group numbers.

0 = root, 102 = polkitd in Arch/Manjaro, polkitd group does not exist in Ubuntu,
so use group numbers instead for the CI's sake,
as both these groups exist by default in UBuntu.
102 is just the systemd-timesync1 group,
but it doesn't matter on the booted Manjaro system.

Signed-off-by: Dan Johansen <strit@manjaro.org>