[buildarmimg] fix rpi4 btrfs bug
There was a bug introduced when we changed the rpi4 kernel packages
to set the PARTUUID's themselves, that resulted in the PARTUUID not
actually being written when using tools and/or installer to install.
It only worked for already installed systems.

So change the sed argument in the tools to take account for this flaw.
And also fix the btrfs cmdline.txt line.

Signed-off-by: Dan Johansen <strit@manjaro.org>